tests/integration/python/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
pytest-xdist==3.3.1
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
kafka-python==2.0.2
redis==5.0.1
prometheus-client==0.19.0
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
pydantic==2.5.2
pydantic-settings==2.1.0
click==8.1.7
colorama==0.4.6
tabulate==0.9.0
psutil==5.9.6
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4